Check for pp archive existence prior to job creation.
Current pp archival cannot cope with data produced by non-current
pp versions. When the version of an artic pp is updated, the
pp archival jobs that are scheduled after the update, but have to
archive data, which was produced prior to the update, fail.

The solution is to check that the pp staging archive for the 'current'
pp version exists. If it does not exist, but there is an archive for
some other version, data from that staging archive should be archived.
Staging archives for multiple pp versions cause the code to error.

Closes #688