Fixed a bug in autoqc checks validation.
Fixed a bug in autoqc checks validation for the npg_run_is_deletable
script. Previously the runs where libraries had the primer panel set
were not autoqc-deletable since the insert_size check results, which
are not run for such products, were absent in the database, but were
expected by the code (closes #710).

Added a new set of test to illustrate both existing and new
functionality for the insert_size check.