Cache process_separately_lanes value and reuse.
If the value of this attrbute is set (in real life scenario via
command line arguments to pipeline scripts), it is cached by the
analysis pipelne only. If the value is not set, an attempt to
retrieve a cached value is made. This ensures that the archival
pipeline and the npg_run_is_deletable script run correctly.