Call _should_run early, make the logic clear.
Changed the signature of _should_run from ($is_plex, $product)
to ($product, $is_lane, $is_plex), so that $is_lane flag is set
by the caller.

Previously it was assumed that anything that is not a plex is
a lane. This is wrong for merged lanes, for which some lane-
level checks cannot run, for example spacial_filter, tag_metrics.